backend/requirements.txt.txt
django
djangorestframework
joblib
django-cors-headers